Extend new event from deprecated event
